#0: Move grad syncing to ttml::core::distributed
