+++ tt-train/sources/examples/nano_gpt/main.cpp
@@ -12,6 +12,7 @@
 
 #include "autograd/tensor.hpp"
 #include "core/clip_grad_norm.hpp"
+#include "core/distributed.hpp"
 #include "core/tt_tensor_utils.hpp"
 #include "datasets/dataloader.hpp"
 #include "datasets/in_memory_token_dataset.hpp"
@@ -635,7 +636,7 @@ int main(int argc, char **argv) {
             gradient_accumulator_helper.update(loss_float, samples);
 
             // synchronize gradients for multi-device case, no-op if single device
-            ttml::optimizers::distributed::synchronize_and_update_grads(model->parameters());
+            ttml::core::distributed::synchronize_parameters(model->parameters());
             if (config.use_clip_grad_norm) {
                 ttml::core::clip_grad_norm(model->parameters(), config.clip_grad_norm_max_norm);
             }
