+++ tt-train/sources/ttml/optimizers/optimizer_base.hpp
@@ -8,13 +8,6 @@
 
 namespace ttml::optimizers {
 
-namespace distributed {
-
-ttnn::Tensor synchronize_tensor(const ttnn::Tensor& tensor);
-void synchronize_and_update_grads(const serialization::NamedParameters& parameters);
-
-}  // namespace distributed
-
 class OptimizerBase {
 public:
     explicit OptimizerBase(serialization::NamedParameters&& parameters);
