+++ tt-train/sources/ttml/optimizers/sgd.cpp
@@ -49,9 +49,6 @@ void SGD::step() {
 
         auto gradients = tensor_ptr->get_grad();
 
-        // synchronize gradients for multi-device case, no-op if single device
-        gradients = distributed::synchronize_tensor(gradients);
-
         if (m_config.weight_decay != 0.0F) {
             gradients = ttnn::add(
                 ttnn::multiply(tensor_ptr->get_value(autograd::PreferredPrecision::FULL), m_config.weight_decay),
